requirements.txt
pyyaml>=6.0
mkdocs>=1.5.0
mkdocs-material>=9.5.0
pymdown-extensions>=10.0
